Update supmapper.py to mirror how SupMover applies target percent
SupMover applies the target percent directly to the luminance value of each pixel in the bitmap rather than in the RGB color space. Previously, the script was calculating the target percent in reference/rgb mode using the RGB color space. This is now fixed to calculate based off the luminance value.

Calculating the target percent in the RGB color space was fairly accurate, but has very rare edge cases where rounding errors could cause the tonemapped PGS file to be 1 RGB value below the reference or specified RGB value.